+++ examples/text/text_font_spritefont.cpp
@@ -54,13 +54,13 @@ int main(void)
         // Draw
         //----------------------------------------------------------------------------------
         BeginDrawing();
-
+        {
             ClearBackground(RAYWHITE);
 
             font1.DrawText(msg1, fontPosition1, font1.baseSize, -3);
             font2.DrawText(msg2, fontPosition2, font2.baseSize, -2);
             font3.DrawText(msg3, fontPosition3, font3.baseSize, 2);
-
+        }
         EndDrawing();
         //----------------------------------------------------------------------------------
     }