+++ examples/textures/textures_image_drawing.cpp
@@ -61,7 +61,7 @@ int main(void)
         // Draw
         //----------------------------------------------------------------------------------
         BeginDrawing();
-
+        {
             background.ClearBackground();
 
             texture.Draw(screenWidth/2 - texture.width/2, screenHeight/2 - texture.height/2 - 40);
@@ -69,7 +69,7 @@ int main(void)
 
             darkGray.DrawText("We are drawing only one texture from various images composed!", 240, 350, 10);
             darkGray.DrawText("Source images have been cropped, scaled, flipped and copied one over the other.", 190, 370, 10);
-
+        }
         EndDrawing();
         //----------------------------------------------------------------------------------
     }