+++ include/Font.hpp
@@ -1,27 +1,3 @@
-/*
-*   LICENSE: zlib/libpng
-*
-*   raylib-cpp is licensed under an unmodified zlib/libpng license, which is an OSI-certified,
-*   BSD-like license that allows static linking with closed source software:
-*
-*   Copyright (c) 2020 Rob Loach (@RobLoach)
-*
-*   This software is provided "as-is", without any express or implied warranty. In no event
-*   will the authors be held liable for any damages arising from the use of this software.
-*
-*   Permission is granted to anyone to use this software for any purpose, including commercial
-*   applications, and to alter it and redistribute it freely, subject to the following restrictions:
-*
-*     1. The origin of this software must not be misrepresented; you must not claim that you
-*     wrote the original software. If you use this software in a product, an acknowledgment
-*     in the product documentation would be appreciated but is not required.
-*
-*     2. Altered source versions must be plainly marked as such, and must not be misrepresented
-*     as being the original software.
-*
-*     3. This notice may not be removed or altered from any source distribution.
-*/
-
 #ifndef RAYLIB_CPP_INCLUDE_FONT_HPP_
 #define RAYLIB_CPP_INCLUDE_FONT_HPP_
 
@@ -31,6 +7,9 @@
 #include "./raylib-cpp-utils.hpp"
 
 namespace raylib {
+/**
+ * Font type, includes texture and charSet array data
+ */
 class Font : public ::Font {
  public:
     Font() {
@@ -79,6 +58,9 @@ class Font : public ::Font {
         return *this;
     }
 
+    /**
+     * Draw text using font and additional parameters.
+     */
     inline Font& DrawText(const std::string& text, ::Vector2 position, float fontSize,
             float spacing, ::Color tint = WHITE) {
         ::DrawTextEx(*this, text.c_str(), position,  fontSize,  spacing,  tint);
@@ -91,6 +73,9 @@ class Font : public ::Font {
         return *this;
     }
 
+    /**
+     * Draw text using font inside rectangle limits with support for text selection.
+     */
     inline Font& DrawText(const std::string& text, ::Rectangle rec, float fontSize, float spacing,
             bool wordWrap, ::Color tint, int selectStart, int selectLength, ::Color selectText,
             ::Color selectBack) {
@@ -132,7 +117,7 @@ class Font : public ::Font {
         return ::ImageTextEx(*this, text.c_str(), fontSize, spacing, tint);
     }
 
- protected:
+ private:
     void set(const ::Font& font) {
         baseSize = font.baseSize;
         charsCount = font.charsCount;