+++ include/physac.hpp
@@ -1,0 +1,15 @@
+/**
+ * C++ header to wrap physac.h.
+ */
+#ifndef RAYLIB_CPP_INCLUDE_PHYSAC_HPP_
+#define RAYLIB_CPP_INCLUDE_PHYSAC_HPP_
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+#include "physac.h" // NOLINT
+#ifdef __cplusplus
+}
+#endif
+
+#endif  // RAYLIB_CPP_INCLUDE_PHYSAC_HPP_