+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.5.0-beta3] - 2021-02-18
+### Added
+- Added a C++ version of raylib's loading thread example (by [@pkeir]([https://github.com/pkeir))
+- Updated documentation
+### Fixed
+- Made the global wrapped functions static to avoid redeclaration
+
 ## [v3.5.0-beta2] - 2021-01-18
 ### Fixed
 - `Font.charsPadding` loading